.backup/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Procurement Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-core",
    "llama-index-readers-file",
    "llama-index-embeddings-huggingface",
    "llama-index-llms-deepseek",
    "uvicorn",
    "fastapi",
    "pydantic-ai",
    "logfire>=4.19.0",
    "torch",
    "python-dotenv",
]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
